Add an auto-modernizer
This is not-quite almost-always-auto, but it is a lot of auto.  Notably,
it does not convert

 QString s;

into

 auto s = QString();

in part because I don't know how to determine with the clang API whether
the type is copyable.

Additionally, while top-scope declarations like

 static const int num = 124;

can in theory be converted to

 static const auto num = 124;

they don't seem to be visited by the VisitStmt API.